+++ mitra_activitypub/src/handlers/offer.rs
@@ -38,8 +38,7 @@ pub async fn handle_offer(
     db_client: &impl DatabaseClient,
     activity: JsonValue,
 ) -> HandlerResult {
-    let activity: Offer = serde_json::from_value(activity)
-        .map_err(|_| ValidationError("unexpected activity structure"))?;
+    let activity: Offer = serde_json::from_value(activity)?;
     let actor_profile = get_remote_profile_by_actor_id(
         db_client,
         &activity.actor,