+++ mitra_activitypub/src/handlers/reject.rs
@@ -34,8 +34,7 @@ pub async fn handle_reject(
     activity: Value,
 ) -> HandlerResult {
     // Reject(Follow)
-    let activity: Reject = serde_json::from_value(activity)
-        .map_err(|_| ValidationError("unexpected activity structure"))?;
+    let activity: Reject = serde_json::from_value(activity)?;
     let actor_profile = get_remote_profile_by_actor_id(
         db_client,
         &activity.actor,